Fix imagePullPolicy override
For development environments, we always want to be running with the latest
images. Given we currently re-use image tags, Kubernetes will by default
not pull the images again, and nothing in the dev tooling would pull them
or indicate you're running with outdated images.

By setting imagePullPolicy to always, we ensure Kubernetes will always
pull the necessary images and keep them up to date as the dev env is
started.