Update documentation to be up-to-date with the most recent changes.
When running `rubocop` is not strictly required to run the `entrypoint.sh`,
but it won't harm (and is easier to remember to run every command in the
container with the `entrypoint.sh`). However, running `rspec` requires
the `entrypoint.sh` to be called, otherwise database credentials won't
be exposed to the process running and it will fail to connect to the
database.